Move files and add `progress` prop
